wip oracle flow

+++ programs/jbl/src/instructions/borrow.rs
@@ -104,7 +104,7 @@ pub fn borrow_handler<'a>(ctx: Context<'a, Borrow<'a>>, amount: u64) -> Result<(
     // ── 1. Accrue interest on the pool via IRM CPI ────────────────────────────
     let (utilization, max_feed_age) = {
         let pool = ctx.accounts.pool.load()?;
-        (pool.calculate_utilization(), pool.max_feed_age_secs as i64)
+        (pool.calculate_utilization(), pool.max_feed_age_secs)
     };
     let oracle = OracleState::new(
         ctx.accounts.feed_program.to_account_info(),
